Bug 1877350 - Refactor tabstray composable to extract tab counter into a separate file
Preparing to use it for navigation bar composable.
